Move ColorDifference struct to its own file
